Проведена работа по задаче 2.a. - получить аватарку пользователя
