--- TickPic/ImagesListCell.swift
@@ -1,17 +1,0 @@
-//
-//  ImagesListCell.swift
-//  TickPic
-//
-//  Created by Кирилл Марьясов on 10.04.2024.
-//
-
-import Foundation
-import UIKit
-
-final class ImagesListCell: UITableViewCell {
-    static let reuseIdentifier = "ImagesListCell"
-    
-    @IBOutlet var cellImage: UIImageView!
-    @IBOutlet var likeButton: UIButton!
-    @IBOutlet var dateLabel: UILabel!
-}
